Add unit test suite
We have sorely needed a unit test suite to work out smaller pieces of
logic in isolation. We should add `greatest` to accomplish this, as it
is a small and self-contained dependency with an ergonomic interface.

We had to move some consts into defines in day2 so they could be reused.