Use relative imports for day header files
We can stop dumping all of src into the compile path if we just use
relative includes in the test files to pull the headers from up a level.
This also fixes the clangd language server, since it now can resolve the
headers and doesn't just give a million errors. Seems like a win-win.